# enhancement(Value)!: CoW value optimization