--- a/examples/simple.rs
+++ b/examples/simple.rs
@@ -1,4 +1,3 @@
-use std::collections::BTreeMap;
 use vrl::{
     compiler::{state::RuntimeState, Context, TargetValue, TimeZone},
     value,