--- a/src/cli/cmd.rs
+++ b/src/cli/cmd.rs
@@ -17,7 +17,7 @@ use crate::compiler::{
 use crate::diagnostic::Formatter;
 use crate::owned_metadata_path;
 use crate::value::Value;
-use crate::value::{ObjectMap, Secrets};
+use crate::value::Secrets;
 use clap::Parser;
 
 use super::repl;