--- a/src/cli/repl.rs
+++ b/src/cli/repl.rs
@@ -5,7 +5,6 @@ use crate::compiler::TimeZone;
 use crate::compiler::{compile_with_state, CompileConfig, Function, Program, Target, VrlRuntime};
 use crate::diagnostic::Formatter;
 use crate::owned_metadata_path;
-use crate::prelude::ObjectMap;
 use crate::value::Secrets;
 use crate::value::Value;
 use indoc::indoc;