--- a/src/compiler/expression/array.rs
+++ b/src/compiler/expression/array.rs
@@ -1,6 +1,6 @@
 use std::{collections::BTreeMap, fmt, ops::Deref};
 
-use crate::value::{ObjectArray, Value};
+use crate::value::{Array as ArrayValue, Value};
 use crate::{
     compiler::{
         expression::{Expr, Resolved},
@@ -51,7 +51,7 @@ impl Expression for Array {
             self.inner
                 .iter()
                 .map(|expr| expr.resolve(ctx))
-                .collect::<Result<ObjectArray, _>>()
+                .collect::<Result<ArrayValue, _>>()
                 .map(Value::Array)
         }
     }
@@ -60,7 +60,7 @@ impl Expression for Array {
         self.inner
             .iter()
             .map(|x| x.resolve_constant(state))
-            .collect::<Option<ObjectArray>>()
+            .collect::<Option<ArrayValue>>()
             .map(Value::Array)
     }