--- a/src/compiler/value.rs
+++ b/src/compiler/value.rs
@@ -3,7 +3,7 @@ mod convert;
 mod error;
 pub mod kind;
 
-pub use crate::value::value::{IterItem, ObjectMap};
+pub use crate::value::value::{IterItem, ObjectMap, ObjectArray};
 #[allow(clippy::module_name_repetitions)]
 pub use error::ValueError;
 pub use kind::{Collection, Field, Index, Kind};