@pront 2025-03-20
Related: https://github.com/vectordotdev/vrl/issues/1021
cc @bruceg you might be interested in this.