@pront 2025-03-20
Hi @Zettroke, thank you for submitting this. I added it to my list for review. Is this a breaking change for users or for developers only?