@Zettroke 2025-03-20
Hi @pront, no it's breaking change only in terms of crate API. VRL itself is unchanged.